Homework 10. Tests

+++ src/main/java/ru/gb/springhw6/repositories/NoteRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface NoteRepository extends JpaRepository<Note, Long> {
-    public Optional<Note> findById(Long id);
+    public Optional<Note> findById(Integer id);
+
 }
